[FE] Feat: 게임 페이지 문제 설명 마크타운 지원
