+++ frontend/package.json
@@ -13,6 +13,7 @@
     "@types/react": "^18.2.36",
     "@types/react-dom": "^18.2.14",
     "@types/react-router-dom": "^5.3.3",
+    "@uiw/react-markdown-preview": "^5.0.3",
     "axios": "^1.6.2",
     "framer-motion": "^10.16.5",
     "korsearch": "^1.1.5",
