+++ frontend/src/pages/GamePlayPage.tsx
@@ -7,7 +7,7 @@ import { getProblemById } from "../api/problem";
 import { motion } from "framer-motion";
 import SlidePage from "../components/common/SlidePage";
 
-const tempProblemId = [1, 2, 4];
+const tempProblemId = [10];
 const GAME_COUNT = 3;
 const GamePlayPage: React.FC = () => {
   const [problemList, setProblemList] = useState<ProblemType[]>([]);
